+++ Utils/Protobuf.h
@@ -70,24 +70,24 @@ public:
 private:
     VarInt Int2Varint(uint64_t l);
     void Int2Varint(uint64_t l, VarInt *VarInt);
-    inline VarInt GetField(uint16_t Field, ProtobufStruct::ProtobufStructType ProtobufStructType) { return Int2Varint(Field << 3 | static_cast<int>(ProtobufStructType)); }
-    inline void GetField(uint16_t Field, ProtobufStruct::ProtobufStructType ProtobufStructType, VarInt &OutField) { Int2Varint(Field << 3 | static_cast<int>(ProtobufStructType), &OutField); }
+    inline VarInt GetField(uint32_t Field, ProtobufStruct::ProtobufStructType ProtobufStructType) { return Int2Varint(Field << 3 | static_cast<int>(ProtobufStructType)); }
+    inline void GetField(uint32_t Field, ProtobufStruct::ProtobufStructType ProtobufStructType, VarInt &OutField) { Int2Varint(Field << 3 | static_cast<int>(ProtobufStructType), &OutField); }
     inline void SetVarint(::Pack *Pack, VarInt *VarInt) { Pack->SetBin_(VarInt->VarInt, VarInt->Length); }
 
     uint32_t Calculate(Tree *Tree);
     void Recurse(::Pack *Pack, Tree *Tree);
 
 public:
-    void WriteVarint(const uint16_t Field, const uint64_t l);
-    void WriteFix32(const uint16_t Field, const int32_t i);
-    void WriteFix64(const uint16_t Field, const int64_t l);
-    void WriteStr(const uint16_t Field, const char8_t *str);
-    void WriteStr_(const uint16_t Field, char8_t *str);
-    void WriteBin(const uint16_t Field, const byte *bin, uint32_t Length);
-    void WriteBin_(const uint16_t Field, byte *bin, uint32_t Length);
-    void WriteBin(const uint16_t Field, const LPBYTE bin);
-    void WriteBin_(const uint16_t Field, const LPBYTE bin);
-    void StepIn(const byte Field);
+    void WriteVarint(const uint32_t Field, const uint64_t l);
+    void WriteFix32(const uint32_t Field, const int32_t i);
+    void WriteFix64(const uint32_t Field, const int64_t l);
+    void WriteStr(const uint32_t Field, const char8_t *str);
+    void WriteStr_(const uint32_t Field, char8_t *str);
+    void WriteBin(const uint32_t Field, const byte *bin, uint32_t Length);
+    void WriteBin_(const uint32_t Field, byte *bin, uint32_t Length);
+    void WriteBin(const uint32_t Field, const LPBYTE bin);
+    void WriteBin_(const uint32_t Field, const LPBYTE bin);
+    void StepIn(const uint32_t Field);
     void StepOut();
     LPBYTE Pack();
     uint32_t Pack(byte *&Bin);
